requirements.txt
alembic==1.4.2
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
click==7.1.2
cssmin==0.2.0
Flask==1.1.2
Flask-Assets==2.0
Flask-FontAwesome==0.1.5
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==5.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
pathspec==0.8.0
pylint==2.6.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
regex==2020.9.27
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
webassets==2.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
